move integration tests to new file
