+++ dynamic/provider_test.go
@@ -13,13 +13,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/hexops/autogold/v2"
-	"github.com/opentofu/opentofu/shim/grpcutil"
-	v6shim "github.com/opentofu/opentofu/shim/protov6"
-	"github.com/pulumi/providertest/pulumitest/optnewstack"
-	"github.com/pulumi/providertest/pulumitest/opttest"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/plugin"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
@@ -31,10 +25,7 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 
 	helper "github.com/pulumi/pulumi-terraform-bridge/dynamic/internal/testing"
-	"github.com/pulumi/pulumi-terraform-bridge/dynamic/parameterize"
-	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/pf/tests/pulcheck"
 	pfbridge "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/pf/tfbridge"
-	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/info"
 )
 
 // globalTempDir is a temporary directory scoped to the entire test cycle.
@@ -554,87 +545,4 @@ func must[T any](v T, err error) T {
 
 func marshal(m resource.PropertyMap) *structpb.Struct {
 	return must(plugin.MarshalProperties(m, plugin.MarshalOptions{}))
-}
-
-func TestLogReplayProvider(t *testing.T) {
-	grpcLogs, err := os.ReadFile("./testdata/TestLogReplayProvider/grpc_log_random.json")
-	if err != nil {
-		t.Fatalf("failed to read grpc log: %v", err)
-	}
-
-	provPlugin := grpcutil.NewLogReplayProvider("random", "0.0.1", string(grpcLogs))
-	prov := v6shim.New(provPlugin)
-	require.NoError(t, err)
-
-	resp, err := prov.GetProviderSchema(context.Background(), &tfprotov6.GetProviderSchemaRequest{})
-	require.NoError(t, err)
-	require.Contains(t, resp.ResourceSchemas, "random_bytes")
-	require.Contains(t, resp.ResourceSchemas, "random_pet")
-	require.Contains(t, resp.ResourceSchemas, "random_string")
-
-	configVal, err := tfprotov6.NewDynamicValue(
-		tftypes.Object{},
-		tftypes.NewValue(tftypes.Object{}, map[string]tftypes.Value{}),
-	)
-	require.NoError(t, err)
-	configResp, err := prov.ValidateProviderConfig(context.Background(), &tfprotov6.ValidateProviderConfigRequest{
-		Config: &configVal,
-	})
-
-	require.NoError(t, err)
-	require.Equal(t, "\x80", string(configResp.PreparedConfig.MsgPack))
-}
-
-type runProvider struct {
-	tfprotov6.ProviderServer
-	name, version string
-}
-
-func (p runProvider) Name() string    { return p.name }
-func (p runProvider) Version() string { return p.version }
-func (p runProvider) URL() string     { return "url" }
-func (p runProvider) Close() error    { return nil }
-
-func makeLogReplayProvider(t *testing.T, name, version string, grpcLogs []byte) info.Provider {
-	provPlugin := grpcutil.NewLogReplayProvider(name, version, string(grpcLogs))
-	prov := v6shim.New(provPlugin)
-	provider := runProvider{
-		ProviderServer: prov,
-		name:           name,
-		version:        version,
-	}
-
-	info, err := providerInfo(context.Background(), provider, parameterize.Value{
-		Local: &parameterize.LocalValue{Path: "path"},
-	})
-	require.NoError(t, err)
-
-	return info
-}
-
-func TestLogReplayProviderWithProgram(t *testing.T) {
-	grpcLogs, err := os.ReadFile(
-		"./testdata/TestLogReplayProvider/grpc_log_random.json")
-	if err != nil {
-		t.Fatalf("failed to read grpc log: %v", err)
-	}
-
-	info := makeLogReplayProvider(t, "random", "0.0.1", grpcLogs)
-	program := `
-name: proj
-runtime: yaml
-resources:
-  randomPet:
-    type: random:Pet
-    properties:
-        length: 3
-outputs:
-  petName: ${randomPet.id}`
-
-	pt, err := pulcheck.PulCheck(t, info, program,
-		opttest.NewStackOptions(optnewstack.DisableAutoDestroy()),
-	)
-	require.NoError(t, err)
-	res := pt.Up(t)
-	require.Equal(t, "heartily-sharing-monkey", res.Outputs["petName"].Value)
 }
